requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
cryptography>=41.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
requests>=2.31.0
PyNaCl>=1.5.0
alembic>=1.13.0
sqlalchemy>=2.0.0